[CLEANUP][swarms.swarms -> swarms.structs]
