+++ swarms/structs/utils.py
