--- swarms/swarms/__init__.py
@@ -1,13 +1,0 @@
-from swarms.structs.autoscaler import AutoScaler
-from swarms.swarms.model_parallizer import ModelParallelizer
-from swarms.swarms.multi_agent_collab import MultiAgentCollaboration
-from swarms.swarms.base import AbstractSwarm
-
-# from swarms.swarms.team import Team
-
-__all__ = [
-    "AutoScaler",
-    "ModelParallelizer",
-    "MultiAgentCollaboration",
-    "AbstractSwarm",
-]
