+++ tests/structs/test_groupchat.py
@@ -3,7 +3,7 @@ import pytest
 from swarms.models import OpenAIChat
 from swarms.models.anthropic import Anthropic
 from swarms.structs.agent import Agent
-from swarms.swarms.groupchat import GroupChat, GroupChatManager
+from swarms.structs.groupchat import GroupChat, GroupChatManager
 
 llm = OpenAIChat()
 llm2 = Anthropic()
